style: separate types
seperate chatroom and chatlist type
